fix:面试日历接入functioncall

+++ eros-db/src/main/java/com/open/ai/eros/db/mysql/hr/entity/IcRecord.java
@@ -81,7 +81,7 @@ public class IcRecord implements Serializable {
     /**
      * 职位id
      */
-    private Integer positionId;
+    private Long positionId;
 
 
 }
